corrected dtype

+++ src/Resampler/loop.h
@@ -4,4 +4,4 @@
 
 void loopSamplerSpecharm(float* input, int length, float* output, int targetLength, float spacing, engineCfg config);
 
-void loopSamplerPitch(short* input, int length, float* output, int targetLength, float spacing);
+void loopSamplerPitch(int* input, int length, float* output, int targetLength, float spacing);
